tmpl8: move cache handling into separate source file
